casts to stop some compiler warnings

+++ src/pl/pl_iter.c
@@ -97,13 +97,13 @@ pv pl_iter_value(pv val) {
 		}
 		switch (i->type) {
 		case KEYS:
-			out = pv_int(i->aiter);
+			out = pv_int((int)i->aiter);
 			break;
 		case VALUES:
 			out = pv_array_get(pv_copy(i->val), i->aiter);
 			break;
 		case ENTRIES:
-			out = PV_ARRAY(pv_int(i->aiter), pv_array_get(pv_copy(i->val), i->aiter));
+			out = PV_ARRAY(pv_int((int)i->aiter), pv_array_get(pv_copy(i->val), i->aiter));
 			break;
 		default:
 			out = pv_invalid(); // just in case
